Rework of the existing Elevator Pitch
Looking at the current Elevator Pitch - it misses the "Extending applications with Kubernetes" angle, which is at the core of the Kyma value proposition and ecosystem. While extension applications are also applications and also deserve a first-class runtime environment to scale, thrive and be operated in - Kyma's unique selling point remains the first-class application connectivity and focus on event-driven scenarios, augmenting new and existing business processes.